NewLLM/requirements.txt
flask==2.3.3
groq==0.4.0
python-dotenv==1.0.0
pydub==0.25.1
numpy==1.26.0
requests==2.31.0 